apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
supabase==2.9.1
pandas==2.2.3
lxml==5.3.0
beautifulsoup4==4.12.3
requests==2.32.3
celery==5.4.0
redis==5.2.0
sentry-sdk[fastapi]==2.17.0
prometheus-fastapi-instrumentator==7.0.0
structlog==24.4.0
openpyxl==3.1.5
xmltodict==0.13.0
ftplib3==0.3.0
paramiko==3.5.0
PyJWT==2.9.0
cryptography==43.0.1